Add comments to the Parse tree, add quotes, add to manual, add test
Comments are now saved in the Parse tree and stripped out by
Intermediate. This prevents some otherwise difficult bugs, like
comments at the very end of a paragraph causing the subsequent blank
line to be a syntax error.

Quotes are now in Markup, parsed and rendered similar to Emph.

The manual now has a section on sections.

Notes were added in comments on the need for the Markup parsers to be
aware of failure with input consumption, so errors can be propagated
better.

A test that Parse+Intermediate is working correctly was added.