Add element reference in manual, lists, fix loop, rename FormalBlock
The manual now has an element reference section, which will contain a
description of each element, along with a brief example of each
(useful for our tests). At the moment it only discusses olist.

We now support simple ordered and unordered lists.

The data FormalBlock is now data Formal, and the FormalBlockBlock
branch of Block is now simply FormalBlock.

A loop was present in the pSpace and pOnlySpace whitespace consumers,
which is now fixed. They didn't update the input state when they
encountered whitespace text nodes, so they would endlessly try to
strip out the same text node. This also meant that pSpace wouldn't
strip off the whitespace of a partially whitespace node.