Add manual comments, fix empty paragraph bug
We need to strip out paragraphs consisting solely of whitespace and
comments during the intermediate conversion, because that's what
comments between blocks in sectioned content look like. We were using
a function to test the input for emptiness, but had designed it as if
it were the output (with comments stripped) that we were testing.

Manual comments relating to macros, templates, numbering, and block
types were added.