Update the README, add the Riemann metadata back to that directory.
The README now has a more detailed discussion of the future. It might
want to go into a separate document, honestly, so that the README is
more friendly.

The Riemann metadata and digitization notes added back to the README
of that document. They still need to be sorted out, but that will have
to wait until a better metadata policy is developed.