Distinguish between white space and non-white space content
This affects both Source and Intermediate, since we would like to
parse white space and non-white-space separately, then pass that
distinction on to Intermediate.

Having this difference around makes certain intermediate parsers
easier to write, notably pOnlySpace, which was written to throw errors
and exclude non-white-space nodes because we didn't have a position
update fold (and indeed one can't be written unless we start storing
the ambient indentation of blocks). Now it can become a simple space
consumer.