Add note text collection to the compiling, add a Bnil branch to Block
The Bnil is necessary because we are removing the note text from the
tree, and we were formerly not allowed to replace it with
anything. This could also be solved by working directly on a list of
blocks (to allow for filtering), but our instances are not really set
up to work like that.